Put main pkg in a sub dir
