+++ .gitignore
@@ -21,5 +21,5 @@ go.work
 .DS_Store
 
 # the produced binary
-vitess-tester
+./vitess-tester
 errors/
